changes after checking

+++ HomeWork_2_5_1/HomeWork_2_5_1/HWMath.cpp
@@ -14,7 +14,7 @@ int multiplication(int num1, int num2) {
 };
 
 double division(int num1, int num2) {
-	return num1 / num2;
+	return static_cast<double>(num1) / num2;
 };
 
 double power(int num1, int num2) {
